// src/app/interfaces/candidato.interface.ts
export interface Candidato{
  key$?:string;
  nome: string;
  sexo: string;
  infantil: string;
  horas: string;
  naturalidade: string;
  uf: string;
  dataNascimento: string;
  identidade: string;
  endereco: string;
  num: string;
  complemento: string;
  bairro: string;
  cep: string;
  telefone: string;
  termo?: string;
  folha?: string;
  livro?: string;
  dataEmissao?: string;
  ufCartorio?: string;
  nomeCartorio?: string;
  temId?: string;
  numId?: string;
  comId?: string;
  dataId?: string;
  ufId?: string;
  orgaoId?: string;
  nomePai?: string;
  nascPai?: string;
  idPai?: string;
  cpfPai?: string;
  instrucaoPai?: string;
  profissaoPai?: string;
  localTrbPai?: string;
  telfPai?: string;
  celularPai?: string;
  emailPai?: string;
  nomeMae?: string;
  nascMae?: string;
  idMae?: string;
  cpfMae?: string;
  instrucaoMae?: string;
  profissaoMae?: string;
  localTrbMae?: string;
  telfMae?: string;
  celularMae?: string;
  emailMae?: string;
  responsavelDoenca?: string;
  necessidadeEspeciais?: string;
  ultimaEscola?: string;
  tratamento?: string;
  qualTratamento?: string;
  alergico?: string;
  medicacao?: string;
  dosagem?: string;
  responsavel?: string;
  nascResponsavel?: string;
  idResponsavel?: string;
  cpfResponsavel?: string;
  instrucaoResponsavel?: string;
  profissaoResponsavel?: string;
  localTrbResp?: string;
  telfResp?: string;
  celularResp?: string;
  emailResp?: string;
  declaraRenda?: string;
  tipoDeclaracao?: string;
  problemaGraveSaude: string;
  remedioControlado: string;
  planoSaude: string;
  alergia: string;
  remedioHabitual: string;
  acidente: string;
  contato1nome: string;
  contato1tel: string;
  contato2nome: string;
  contato2tel: string;
  aceita: string;
}